+++ src/assets/scripts/client/ui/UiController.js
@@ -5,6 +5,7 @@ import EventBus from '../lib/EventBus';
 import EventTracker from '../EventTracker';
 import GameController from '../game/GameController';
 import SettingsController from './SettingsController';
+import TrafficRateController from './TrafficRateController';
 import TutorialView from './TutorialView';
 import { speech_toggle } from '../speech';
 import { EVENT } from '../constants/eventNames';
@@ -40,6 +41,7 @@ class UiController {
         this.$toggleTutorial = null;
         this.$toggleOptions = null;
         this.$toggleVideoMap = null;
+        this.$toggleTraffic = null;
     }
 
     /**
@@ -55,6 +57,7 @@ class UiController {
         this._eventBus = EventBus;
         this.tutorialView = new TutorialView($element);
         this.settingsController = new SettingsController($element);
+        this.trafficRateController = new TrafficRateController($element);
 
         this.$element = $element;
         this.$airportDialog = this.$element.find(SELECTORS.DOM_SELECTORS.AIRPORT_SWITCH);
@@ -73,6 +76,7 @@ class UiController {
         this.$toggleTutorial = this.$element.find(SELECTORS.DOM_SELECTORS.TOGGLE_TUTORIAL);
         this.$toggleOptions = this.$element.find(SELECTORS.DOM_SELECTORS.TOGGLE_OPTIONS);
         this.$toggleVideoMap = this.$element.find(SELECTORS.DOM_SELECTORS.TOGGLE_VIDEO_MAP);
+        this.$toggleTraffic = this.$element.find(SELECTORS.DOM_SELECTORS.TOGGLE_TRAFFIC);
 
         return this.setupHandlers()
             .enable();
@@ -111,6 +115,7 @@ class UiController {
         this.$toggleTutorial.on('click', (event) => this.onToggleTutorial(event));
         this.$toggleOptions.on('click', (event) => this.onToggleOptions(event));
         this.$toggleVideoMap.on('click', (event) => this.onToggleVideoMap(event));
+        this.$toggleTraffic.on('click', (event) => this.onToggleTraffic(event));
 
         return this;
     }
@@ -136,6 +141,7 @@ class UiController {
         this.$toggleTutorial.off('click', (event) => this.onToggleTutorial(event));
         this.$toggleOptions.off('click', (event) => this.onToggleOptions(event));
         this.$toggleVideoMap.off('click', (event) => this.onToggleVideoMap(event));
+        this.$toggleTraffic.off('click', (event) => this.onToggleTraffic(event));
 
         return this.destroy();
     }
@@ -168,6 +174,7 @@ class UiController {
         this.$toggleTutorial = null;
         this.$toggleOptions = null;
         this.$toggleVideoMap = null;
+        this.$onToggleTraffic = null;
 
         return this;
     }
@@ -179,8 +186,6 @@ class UiController {
     ui_init() {
         this.tutorialView.tutorial_init_pre();
         this.$fastForwards.prop('title', 'Set time warp to 2');
-        // TODO: Make the options dialog findable by ID, not just by class
-        this.$optionsDialog = this.$element.find(SELECTORS.DOM_SELECTORS.OPTIONS_DIALOG);
     }
 
     /**
@@ -194,8 +199,12 @@ class UiController {
             this.tutorialView.tutorial_close();
         }
 
-        if (this.isOptionsDialogOpen()) {
+        if (this.settingsController.isDialogOpen()) {
             this.onToggleOptions();
+        }
+
+        if (this.trafficRateController.isDialogOpen()) {
+            this.onToggleTraffic();
         }
 
         if (this.isAirportSelectionDialogOpen()) {
@@ -212,17 +221,6 @@ class UiController {
      */
     isAirportSelectionDialogOpen() {
         return this.$airportDialog.hasClass(SELECTORS.CLASSNAMES.OPEN);
-    }
-
-    /**
-     * Returns whether the airport selection dialog is open
-     *
-     * @for UiController
-     * @method isOptionsDialogOpen
-     * @return {boolean}
-     */
-    isOptionsDialogOpen() {
-        return this.$optionsDialog.hasClass(SELECTORS.CLASSNAMES.OPEN);
     }
 
     /**
@@ -449,7 +447,7 @@ class UiController {
     onToggleOptions() {
         EventTracker.recordEvent(TRACKABLE_EVENT.SETTINGS, 'toggle-dialog', `${this.$toggleOptions.hasClass(SELECTORS.CLASSNAMES.ACTIVE)}`);
         this.$toggleOptions.toggleClass(SELECTORS.CLASSNAMES.ACTIVE);
-        this.$optionsDialog.toggleClass(SELECTORS.CLASSNAMES.OPEN);
+        this.settingsController.toggleDialog();
     }
 
     /**
@@ -526,6 +524,18 @@ class UiController {
      */
     onClickGithubLink(event) {
         EventTracker.recordClickOnOutboundLink(event.target.href);
+    }
+
+    /**
+     * Handler for toggling the traffic volume view
+     *
+     * @for UiController
+     * @method onToggleTraffic
+     * @param event {jquery event}
+     */
+    onToggleTraffic(event) {
+        this.$toggleTraffic.toggleClass(SELECTORS.CLASSNAMES.ACTIVE);
+        this.trafficRateController.toggleDialog();
     }
 }